requirements.txt
albumentations==1.1.0
async-generator==1.10
colorama==0.4.4
cycler==0.11.0
efficientnet-pytorch==0.6.3
fonttools==4.28.5
imageio==2.13.5
ipython-genutils==0.2.0
joblib==1.1.0
kiwisolver==1.3.2
matplotlib==3.5.1
munch==2.5.0
networkx==2.6.3
numpy==1.22.0
opencv-python==4.5.5.62
opencv-python-headless==4.5.5.62
packaging==21.3
Pillow==9.0.0
pretrainedmodels==0.7.4
pyparsing==3.0.6
PyQt5==5.12.3
PyQt5_sip==4.19.18
PyQtChart==5.12
PyQtWebEngine==5.12.1
python-dateutil==2.8.2
PyWavelets==1.2.0
pywin32==303
PyYAML==6.0
qudida==0.0.4
scikit-image==0.19.1
scikit-learn==1.0.2
scipy==1.7.3
segmentation-models-pytorch==0.2.1
six==1.16.0
threadpoolctl==3.0.0
timm==0.4.12
tqdm==4.62.3
typing_extensions==4.0.1
webencodings==0.5.1